% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_group_terms.R
\name{clust_group_terms}
\alias{clust_group_terms}
\title{clust_group_terms}
\usage{
clust_group_terms(cg)
}
\arguments{
\item{cg}{A data frame output from the GSEAmining cluster_groups function.}
}
\value{

}
\description{
Takes the output of clust_groups, a data frame , and process
it to obtain the enrichment of terms in gene sets names within each cluster.
The output is used in the functions gm_enrichterms and gm_enrichreport.
}
