% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_groups.R
\name{clust_groups}
\alias{clust_groups}
\title{clust_groups}
\usage{
clust_groups(df, hc)
}
\arguments{
\item{df}{Data frame that contains at least an ID column for the gene set
names and a core_enrichment column containing the genes in the leading edge
of each gene set separated by '/'.}

\item{hc}{The output of gm_clust, which is an hclust class object.}
}
\value{

}
\description{
Takes the output of gm_clust, which is an hclust class object,
and returns a data frame that will be used in the rest of GSEAmining
functions gm_enrichreport, gm_enrichterms and gm_enrichcores.
}
