% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{gm_clust}
\alias{gm_clust}
\title{gm_clust: GSEAmining cluster object}
\usage{
gm_clust(df)
}
\arguments{
\item{df}{Data frame that contains at least an ID column for the gene set
names and a core_enrichment column containing the genes in the leading edge
of each gene set separated by '/'.}
}
\value{
A hclust class object with the clustering of the gene sets
 by the core enriched genes.
}
\description{
From the results of GSEA analysis, a data frame:
1. Tokenizes the core-enrichment column.
2. Cast the new table into a matrix.
3. Returns and hclust object.
}
