% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\name{stop_words}
\alias{stop_words}
\title{Stop words.
Eliminates the first word of the gene sets from MSigDb that relate to the
origin of the gene set. Additionally it eliminates words that do not add
a lot of significance such as prepositions or adverbs among others.}
\usage{
stop_words()
}
\description{
#' @return
}
