% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{gm_clust}
\alias{gm_clust}
\title{gm_clust: GSEAmining cluster object}
\usage{
gm_clust(df)
}
\arguments{
\item{df}{Data frame that contains at least three columns: an ID column for
the gene set names, a NES column with the normalized enrichment score and a
core_enrichment column containing the genes in the leading edge of each
gene set separated by '/'.}
}
\value{
An object of class hclust that contains the clustering of the gene
sets by the core enriched genes.First a distance matrix is calculated
using the 'binary' method and then a cluster with the 'complete' method is
created.
}
\description{
Takes the output of gm_filter or a data frame that with the
results of GSEA analysis and returns and hclust object that can be plotted
using the gm_dendplot function.
}
\examples{
data(genesets_sel)
gs.cl <- gm_clust(genesets_sel)


}
