% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_dendplot.R
\name{gm_dendplot}
\alias{gm_dendplot}
\title{gm_dendplot: GSEAmining dendrogram plotter}
\usage{
gm_dendplot(df, hc, col_pos = "red", col_neg = "blue", rect = TRUE)
}
\arguments{
\item{df}{Data frame that contains at least an ID column for the gene set
names and a core_enrichment column containing the genes in the leading edge
of each gene set separated by '/'.}

\item{hc}{The output of gm_clust, which is an hclust class object.}

\item{col_pos}{Color to represent the positively enriched gene sets. Default
is red.}

\item{col_neg}{Color to represent the negatively enriched gene sets. Default
is blue.}

\item{rect}{A logical value indicating if rectangles should be drawn around
the clusters to help differentiating them.}
}
\value{

}
\description{
Takes the output of gm_clust, which is an hclust class object,
and plots the dendrogram using the dendextend package.
}
